pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "クラスタ内 A2A エージェントを束ねる Supervisor 型 ReAct エージェント"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.14.0",
    "openai>=1.54.0",
    "chainlit>=2.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.4.0",
    "pyyaml>=6.0.1",
    "tenacity>=9.0.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "langfuse>=2.50.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.1",
    "pytest-httpserver>=1.1.0",
    "ruff>=0.6.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
